Rename virtual clients name
