+++ src/App.js
@@ -6,10 +6,9 @@ import i18n from "i18next";
 import "./i18n/i18n";
 // import GalaxyApp from "./apps/GalaxyApp";
 // import MobileClient from "./apps/MobileApp/MobileClient";
-// import VirtualClient from "./apps/VirtualApp/VirtualClient";
 // import GalaxyStream from "./apps/StreamApp/GalaxyStream";
-import GalaxyClient from "./apps/VirtualApp/GalaxyClient";
-// import VirtualClient from "./apps/VirtualApp/VirtualClient";
+import GalaxyMqttClient from "./apps/VirtualApp/VirtualMqttClient";
+// import VirtualHttpClient from "./apps/VirtualApp/VirtualHttpClient";
 // import AdminApp from "./apps/AdminApp/AdminApp";
 // import ShidurApp from "./apps/ShidurApp/ShidurApp";
 // import SndmanApp from "./apps/SndmanApp/SndmanApp";
@@ -23,10 +22,9 @@ class App extends Component {
       <I18nextProvider i18n={i18n}>
         {/*{<GalaxyApp />}*/}
         {/*<MobileClient />*/}
-        {/*<VirtualClient />*/}
         {/*<GalaxyStream />*/}
-        <GalaxyClient />
-        {/*<VirtualClient />*/}
+        <GalaxyMqttClient />
+        {/*<VirtualHttpClient />*/}
         {/*<AdminApp />*/}
         {/*<ShidurApp />*/}
         {/*<SndmanApp />*/}
