+++ src/apps/VirtualApp/VirtualMqttClient.js
@@ -1,5 +1,4 @@
 import React, {Component, Fragment} from "react";
-import {Janus} from "../../lib/janus";
 import classNames from "classnames";
 import {isMobile} from "react-device-detect";
 import {Button, Icon, Image, Input, Label, Menu, Message, Modal, Popup, Select} from "semantic-ui-react";
@@ -9,7 +8,6 @@ import {
   getDateString,
   getMedia,
   getMediaStream,
-  initJanus,
   micLevel,
   notifyMe,
   testMic,
@@ -36,7 +34,7 @@ import {
 } from "../../shared/MonitoringData";
 import api from "../../shared/Api";
 import VirtualStreaming from "./VirtualStreaming";
-import VirtualStreamingJanus from "../../shared/VirtualStreamingJanus";
+import JanusStream from "../../shared/streaming-utils";
 import {getUser, kc} from "../../components/UserManager";
 import LoginPage from "../../components/LoginPage";
 import {Profile} from "../../components/Profile";
@@ -46,11 +44,9 @@ import audioModeSvg from "../../shared/audio-mode.svg";
 import fullModeSvg from "../../shared/full-mode.svg";
 import ConfigStore from "../../shared/ConfigStore";
 import {toggleFullScreen, isFullScreen} from "./FullScreenHelper";
-
 import {AppBar, Badge, Box, Button as ButtonMD, ButtonGroup, Grid, IconButton} from "@material-ui/core";
 import {ChevronLeft, ChevronRight, PlayCircleOutline /*, OpenInNewOutlined*/} from "@material-ui/icons"; //button of congress
 import {grey} from "@material-ui/core/colors";
-
 import {AskQuestion, AudioMode, CloseBroadcast, Layout, Mute, MuteVideo, Vote, Fullscreen} from "./buttons";
 import Settings from "./settings/Settings";
 import SettingsJoined from "./settings/SettingsJoined";
@@ -59,15 +55,17 @@ import {SupportOld, Support, initCrisp} from "./components/Support";
 import SendQuestionContainer from "./components/SendQuestions/container";
 import {RegistrationModals} from "./components/RegistrationModals";
 import {getUserRole, userRolesEnum} from "../../shared/enums";
-import KliOlamiStream from "./components/KliOlamiStream";
-import {iceRestart as iceRestartKliOlami} from "./components/KliOlamiStreamHelper";
+import QuadStream from "./components/QuadStream";
+//import {iceRestart as iceRestartKliOlami} from "./components/KliOlamiStreamHelper";
 import KliOlamiToggle from "./buttons/KliOlamiToggle";
 import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
-
 import {withTheme} from "@material-ui/core/styles";
 import ThemeSwitcher from "./components/ThemeSwitcher/ThemeSwitcher";
 import mqtt from "../../shared/mqtt";
+import {JanusMqtt} from "../../lib/janus-mqtt";
+import {PublisherPlugin} from "../../lib/publisher-plugin";
+import {SubscriberPlugin} from "../../lib/subscriber-plugin";
 
 const toggleDesignVersions = () => {
   window.location = isUseNewDesign ? "https://galaxy.kli.one/user/" : "https://arvut.kli.one/user/";
@@ -91,7 +89,7 @@ const userFeeds = (feeds) => feeds.filter((feed) => feed.display.role === userRo
 const isUseNewDesign = true;
 //const isUseNewDesign = /arvut/.test(window.location.host);
 
-class VirtualClient extends Component {
+class VirtualMqttClient extends Component {
   state = {
     chatMessagesCount: 0,
     creatingFeed: false,
@@ -143,7 +141,7 @@ class VirtualClient extends Component {
     currentLayout: localStorage.getItem("currentLayout") || "split",
     attachedSource: true,
     sourceLoading: true,
-    virtualStreamingJanus: new VirtualStreamingJanus(() => this.virtualStreamingInitialized()),
+    virtualStreamingJanus: new JanusStream(() => this.virtualStreamingInitialized()),
     appInitError: null,
     upval: null,
     net_status: 1,
@@ -214,7 +212,6 @@ class VirtualClient extends Component {
 
   checkPermission = (user) => {
     user.role = getUserRole();
-
     if (user.role !== null) {
       this.initApp(user);
     } else {
@@ -230,29 +227,29 @@ class VirtualClient extends Component {
     this.initMQTT(user);
 
     //Clients not authorized to app may see shidur only
-    if (user.role !== userRolesEnum.user) {
-      const config = {
-        gateways: {
-          streaming: {
-            str: {
-              name: "str",
-              url: APP_JANUS_SRV_STR1,
-              type: "streaming",
-              token: "",
-            },
-          },
-        },
-        ice_servers: {streaming: [APP_STUN_SRV_STR]},
-        dynamic_config: {galaxy_premod: "false"},
-        last_modified: new Date().toISOString(),
-      };
-      ConfigStore.setGlobalConfig(config);
-      GxyJanus.setGlobalConfig(config);
-      localStorage.setItem("room", "-1");
-      this.state.virtualStreamingJanus.init(user);
-      this.setState({user, sourceLoading: true});
-      return;
-    }
+    // if (user.role !== userRolesEnum.user) {
+    //   const config = {
+    //     gateways: {
+    //       streaming: {
+    //         str: {
+    //           name: "str",
+    //           url: APP_JANUS_SRV_STR1,
+    //           type: "streaming",
+    //           token: "",
+    //         },
+    //       },
+    //     },
+    //     ice_servers: {streaming: [APP_STUN_SRV_STR]},
+    //     dynamic_config: {galaxy_premod: "false"},
+    //     last_modified: new Date().toISOString(),
+    //   };
+    //   ConfigStore.setGlobalConfig(config);
+    //   GxyJanus.setGlobalConfig(config);
+    //   localStorage.setItem("room", "-1");
+    //   this.state.virtualStreamingJanus.init("", "IL");
+    //   this.setState({user, sourceLoading: true});
+    //   return;
+    // }
 
     const {t} = this.props;
     localStorage.setItem("question", false);
@@ -376,32 +373,63 @@ class VirtualClient extends Component {
     }
 
     const config = GxyJanus.instanceConfig(user.janus);
-    initJanus(
-      (janus) => {
-        // Check if unified plan supported
-        if (Janus.unifiedPlan) {
-          user.session = janus.getSessionId();
-          this.setState({janus});
-          this.initVideoRoom(reconnect, user);
-        } else {
-          alert(t("oldClient.unifiedPlanNotSupported"));
-        }
-      },
-      (err) => {
-        this.exitRoom(/* reconnect= */ true, () => {
-          console.error("[User] error initializing janus", err);
-          this.reinitClient(retry);
-        });
-      },
-      config.url,
-      config.token,
-      config.iceServers
-    );
+    console.log(config)
+    this.initJanus(user, config)
+    // initJanus(
+    //   (janus) => {
+    //     // Check if unified plan supported
+    //     if (Janus.unifiedPlan) {
+    //       user.session = janus.getSessionId();
+    //       this.setState({janus});
+    //       this.initVideoRoom(reconnect, user);
+    //     } else {
+    //       alert(t("oldClient.unifiedPlanNotSupported"));
+    //     }
+    //   },
+    //   (err) => {
+    //     this.exitRoom(/* reconnect= */ true, () => {
+    //       console.error("[User] error initializing janus", err);
+    //       this.reinitClient(retry);
+    //     });
+    //   },
+    //   config.url,
+    //   config.token,
+    //   config.iceServers
+    // );
 
     if (!reconnect) {
       this.state.virtualStreamingJanus.init(user);
     }
   };
+
+  initJanus = (user, config) => {
+    let janus = new JanusMqtt(user, config.name, "MqttGalaxy")
+
+    let videoroom = new PublisherPlugin();
+    videoroom.subTo = this.makeSubscription;
+    videoroom.unsubFrom = this.unsubscribeFrom
+    videoroom.talkEvent = this.handleTalking
+
+    let subscriber = new SubscriberPlugin();
+    subscriber.onTrack = this.onRemoteTrack;
+    subscriber.onUpdate = this.onUpdateStreams;
+
+    janus.init(config.token).then(data => {
+      console.log("[client] Janus init", data)
+
+      janus.attach(videoroom).then(data => {
+        this.setState({janus, videoroom, user});
+        console.log('[client] Publisher Handle: ', data)
+        this.joinRoom(false, videoroom, user)
+      })
+
+      janus.attach(subscriber).then(data => {
+        this.setState({subscriber});
+        console.log('[client] Subscriber Handle: ', data)
+      })
+
+    })
+  }
 
   reinitClient = (retry) => {
     retry++;
@@ -476,7 +504,8 @@ class VirtualClient extends Component {
   setVideoSize = (video_setting) => {
     let {media} = this.state;
     if (JSON.stringify(video_setting) === JSON.stringify(media.video.setting)) return;
-    getMediaStream(false, true, video_setting, null, media.video.video_device).then((data) => {
+    getMediaStream(false, true, video_setting, null, media.video.video_device)
+      .then((data) => {
       console.log(data);
       const [stream, error] = data;
       if (error) {
@@ -495,7 +524,8 @@ class VirtualClient extends Component {
   setVideoDevice = (video_device, reconnect) => {
     let {media} = this.state;
     if (video_device === media.video.video_device && !reconnect) return;
-    return getMediaStream(false, true, media.video.setting, null, video_device).then((data) => {
+    return getMediaStream(false, true, media.video.setting, null, video_device)
+      .then((data) => {
       console.log(data);
       const [stream, error] = data;
       if (error) {
@@ -514,7 +544,8 @@ class VirtualClient extends Component {
   setAudioDevice = (audio_device) => {
     let {media} = this.state;
     if (audio_device === media.audio.audio_device) return;
-    getMediaStream(true, false, media.video.setting, audio_device, null).then((data) => {
+    getMediaStream(true, false, media.video.setting, audio_device, null)
+      .then((data) => {
       console.log(data);
       const [stream, error] = data;
       if (error) {
@@ -649,114 +680,52 @@ class VirtualClient extends Component {
     }
   };
 
-  initVideoRoom = (reconnect, user) => {
-    this.state.janus.attach({
-      plugin: "janus.plugin.videoroom",
-      opaqueId: user.id,
-      success: (videoroom) => {
-        Janus.log(" :: My handle: ", videoroom);
-        Janus.log("Plugin attached! (" + videoroom.getPlugin() + ", id=" + videoroom.getId() + ")");
-        Janus.log("  -- This is a publisher/manager");
-        user.handle = videoroom.getId(); // User state updated in this.joinRoom.
-        this.setState({videoroom});
-        this.joinRoom(reconnect, videoroom, user);
-      },
-      error: (error) => {
-        Janus.log("Error attaching plugin: " + error);
-      },
-      consentDialog: (on) => {
-        Janus.debug("Consent dialog should be " + (on ? "on" : "off") + " now");
-      },
-      iceState: (state) => {
-        Janus.log("ICE state changed to " + state);
-        this.setState({ice: state});
-        this.state.monitoringData.onIceState(state);
-        if (state === "disconnected") {
-          // Chrome: iceconnectionstate does not go to failed if connection drops - https://bugs.chromium.org/p/chromium/issues/detail?id=982793
-          // Safari/Firefox ice restart may be triggered on state: failed as in example - https://developer.mozilla.org/en-US/docs/Web/API/RTCPeerConnection/restartIce
-          this.iceState();
-        }
-      },
-      mediaState: (media, on) => {
-        Janus.log("Janus " + (on ? "started" : "stopped") + " receiving our " + media);
-        this.setState({[media]: on});
-        if (!on) {
-          this.mediaState(media);
-        }
-      },
-      webrtcState: (on) => {
-        Janus.log("Janus says our WebRTC PeerConnection is " + (on ? "up" : "down") + " now");
-      },
-      slowLink: (uplink, lost, mid) => {
-        const slowLinkType = uplink ? "sending" : "receiving";
-        Janus.log("Janus reports problems " + slowLinkType + " packets on mid " + mid + " (" + lost + " lost packets)");
-        this.state.monitoringData.onSlowLink(slowLinkType, lost);
-      },
-      onmessage: (msg, jsep) => {
-        this.onMessage(this.state.videoroom, msg, jsep, false);
-      },
-      onlocaltrack: (track, on) => {
-        Janus.log(" ::: Got a local track event :::");
-        Janus.log("Local track " + (on ? "added" : "removed") + ":", track);
-        let {videoroom} = this.state;
-        videoroom.muteAudio();
-        if (track && track.kind) {
-          if (track.kind === "video") {
-            const localVideoTrack = on ? track : null;
-            this.setState({localVideoTrack});
-          }
-          if (track.kind === "audio") {
-            const localAudioTrack = on ? track : null;
-            this.setState({localAudioTrack});
-          }
-        }
-      },
-      onremotestream: (stream) => {
-        // The publisher stream is sendonly, we don't expect anything here
-        Janus.warn("Send only publisher stream, if this happends, it is not expected, stream:", stream);
-      },
-      ondataopen: (label) => {
-        Janus.log("Publisher - DataChannel is available! (" + label + ")");
-      },
-      ondata: (data, label) => {
-        Janus.log("Publisher - Got data from the DataChannel! (" + label + ")" + data);
-      },
-      ondataerror: (error) => {
-        Janus.warn("Publisher - DataChannel error: " + error);
-      },
-      oncleanup: () => {
-        Janus.log(" ::: Got a cleanup notification: we are unpublished now :::");
-      },
-    });
-  };
-
   joinRoom = (reconnect, videoroom, user) => {
-    let {selected_room, tested, media, cammuted} = this.state;
-    const {
-      video: {video_device},
-    } = media;
+    let {selected_room, tested, media, cammuted, janus} = this.state;
+    const {video: {video_device}} = media;
     user.camera = !!video_device && cammuted === false;
     user.self_test = tested;
     user.sound_test = reconnect ? JSON.parse(localStorage.getItem("sound_test")) : false;
     user.question = false;
     user.timestamp = Date.now();
-    this.setState({user, muted: true});
+    user.session = janus.sessionId;
+    user.handle = videoroom.janusHandleId;
+
+    this.micMute()
+
+    this.setState({user});
 
     updateSentryUser(user);
 
     const {id, timestamp, role, username} = user;
     const d = {id, timestamp, role, display: username};
-    const register = {request: "join", room: selected_room, ptype: "publisher", display: JSON.stringify(d)};
-    videoroom.send({
-      message: register,
-      success: () => {
-        console.log("Request join success");
-      },
-      error: (error) => {
-        console.error(error);
+
+    videoroom.join(selected_room, d).then(data => {
+      console.log('[client] Joined respond :', data)
+      const {audio, video} = this.state.media;
+      videoroom.publish(video.stream, audio.stream)
+
+      const {id, private_id, room} = data
+      user.rfid = data.id;
+      this.setState({user, myid: id, mypvtid: private_id, room, delay: false, wipSettings: false});
+      updateSentryUser(user);
+      updateGxyUser(user);
+      this.keepAlive();
+
+      mqtt.join("galaxy/room/" + selected_room);
+      mqtt.join("galaxy/room/" + selected_room + "/chat", true);
+
+      const feeds = sortAndFilterFeeds(data.publishers.filter((l) => (l.display = JSON.parse(l.display))));
+      console.log("[client] Pulbishers list: ", feeds);
+      // Feeds count with user role
+      let feeds_count = userFeeds(feeds).length;
+      if (feeds_count > 25) {
+        alert(t("oldClient.maxUsersInRoom"));
         this.exitRoom(/* reconnect= */ false);
-      },
-    });
+      }
+      this.makeSubscription(feeds);
+
+    })
   };
 
   exitRoom = (reconnect, callback, error) => {
@@ -777,376 +746,123 @@ class VirtualClient extends Component {
         console.error("Error exiting room", err);
       });
 
-    let {videoroom, remoteFeed, janus, room, shidur, virtualStreamingJanus} = this.state;
-    if (remoteFeed) remoteFeed.detach();
-    if (videoroom) videoroom.send({message: {request: "leave", room}});
+    let {videoroom, subscriber, janus, room, shidur, virtualStreamingJanus} = this.state;
+    // if (remoteFeed) remoteFeed.detach();
+    // if (videoroom) videoroom.send({message: {request: "leave", room}});
 
-    mqtt.exit("galaxy/room/" + room);
-    mqtt.exit("galaxy/room/" + room + "/chat");
+    videoroom.leave().then(data => {
+      console.log("[client] leave respond:", data);
 
-    if (shidur && !reconnect) {
-      virtualStreamingJanus.destroy({
-        success: () => {
-          console.log("Virtual streaming destroyed on exit room.");
-        },
-        error: (error) => {
-          console.log("Error destroying VirtualStreaming on exit room", error);
-        },
-      });
-    }
-    if (!reconnect && isFullScreen()) {
-      toggleFullScreen();
-    }
-    setTimeout(() => {
-      if (videoroom) videoroom.detach();
-      if (janus) janus.destroy();
-      if (!reconnect) {
-        this.state.virtualStreamingJanus.muteAudioElement();
-      } else {
-        this.state.virtualStreamingJanus.unmuteAudioElement();
-      }
-      this.setState({
-        muted: false,
-        question: false,
-        feeds: [],
-        mids: [],
-        localAudioTrack: null,
-        localVideoTrack: null,
-        upval: null,
-        remoteFeed: null,
-        videoroom: null,
-        janus: null,
-        delay: reconnect,
-        room: reconnect ? room : "",
-        chatMessagesCount: 0,
-        isSettings: false,
-      });
-      if (typeof callback === "function") callback();
-    }, 2000);
-  };
+      janus.destroy()
 
-  publishOwnFeed = (useVideo, useAudio) => {
-    const {videoroom, media} = this.state;
-    const {
-      audio: {audio_device},
-      video: {setting, video_device},
-    } = media;
-    const offer = {audioRecv: false, videoRecv: false, audioSend: useAudio, videoSend: useVideo, data: false};
+      mqtt.exit("galaxy/room/" + room);
+      mqtt.exit("galaxy/room/" + room + "/chat");
 
-    if (useVideo) {
-      const {width, height, ideal} = setting;
-      offer.video = {width, height, frameRate: {ideal, min: 1}, deviceId: {exact: video_device}};
-    }
-
-    if (useAudio) {
-      offer.audio = {noiseSuppression: true, deviceId: {exact: audio_device}};
-    }
-
-    videoroom.createOffer({
-      media: offer,
-      simulcast: false,
-      success: (jsep) => {
-        Janus.debug("Got publisher SDP!");
-        Janus.debug(jsep);
-        const publish = {request: "configure", audio: useAudio, video: useVideo, data: false};
-        videoroom.send({message: publish, jsep: jsep});
-        if (!useVideo) {
-          media.video?.stream?.getTracks().forEach((t) => t.stop());
-          this.setState({cammuted: true});
-        }
-      },
-      error: (error) => {
-        Janus.error("WebRTC error:", error);
-      },
-    });
-  };
-
-  iceRestart = () => {
-    const {videoroom, remoteFeed} = this.state;
-
-    if (videoroom) {
-      videoroom.createOffer({
-        media: {audioRecv: false, videoRecv: false, audioSend: true, videoSend: true},
-        iceRestart: true,
-        simulcast: false,
-        success: (jsep) => {
-          Janus.debug("Got publisher SDP!");
-          Janus.debug(jsep);
-          const publish = {request: "configure", restart: true};
-          videoroom.send({message: publish, jsep: jsep});
-        },
-        error: (err) => {
-          Janus.error("WebRTC error:", err);
-        },
-      });
-    }
-
-    if (remoteFeed) remoteFeed.send({message: {request: "configure", restart: true}});
-    if (this.state.virtualStreamingJanus) this.state.virtualStreamingJanus.iceRestart();
-    iceRestartKliOlami();
-  };
-
-  onMessage = (videoroom, msg, jsep) => {
-    const {t} = this.props;
-    Janus.log(`::: Got a message (publisher) :::`, msg);
-    const event = msg["videoroom"];
-    if (event !== undefined && event !== null) {
-      if (event === "joined") {
-        const user = Object.assign({}, this.state.user);
-        const myid = msg["id"];
-        let mypvtid = msg["private_id"];
-        Janus.log("Successfully joined room " + msg["room"] + " with ID " + myid);
-
-        user.rfid = myid;
-        this.setState({user, myid, mypvtid, room: msg["room"], delay: false, wipSettings: false});
-
-        updateSentryUser(user);
-        updateGxyUser(user);
-
-        this.keepAlive();
-
-        // Subscribe to mqtt topic
-        // FIXME: Make sure here the stream is initialized
-        setTimeout(() => {
-          mqtt.join("galaxy/room/" + msg["room"]);
-          mqtt.join("galaxy/room/" + msg["room"] + "/chat", true);
-        }, 3000);
-
-        const {
-          media: {
-            audio: {audio_device},
-            video: {video_device},
+      if (shidur && !reconnect) {
+        virtualStreamingJanus.destroy({
+          success: () => {
+            console.log("Virtual streaming destroyed on exit room.");
           },
-          cammuted,
-        } = this.state;
-        this.publishOwnFeed(!!video_device && !cammuted, !!audio_device);
-
-        // Any new feed to attach to?
-        if (msg["publishers"] !== undefined && msg["publishers"] !== null) {
-          //FIXME: display property is JSON write now, let parse it in one place
-          const feeds = sortAndFilterFeeds(msg["publishers"].filter((l) => (l.display = JSON.parse(l.display))));
-          Janus.log(":: Got Pulbishers list: ", feeds);
-
-          // Feeds count with user role
-          let feeds_count = userFeeds(feeds).length;
-          if (feeds_count > 25) {
-            alert(t("oldClient.maxUsersInRoom"));
-            this.exitRoom(/* reconnect= */ false);
-          }
-
-          Janus.debug("Got a list of available publishers/feeds:");
-          Janus.log(feeds);
-          this.makeSubscription(feeds);
-        }
-      } else if (event === "talking") {
-        const feeds = Object.assign([], this.state.feeds);
-        const id = msg["id"];
-        Janus.log("User: " + id + " - start talking");
-        for (let i = 0; i < feeds.length; i++) {
-          if (feeds[i] && feeds[i].id === id) {
-            feeds[i].talking = true;
-          }
-        }
-        this.setState({feeds});
-      } else if (event === "stopped-talking") {
-        const feeds = Object.assign([], this.state.feeds);
-        const id = msg["id"];
-        Janus.log("User: " + id + " - stop talking");
-        for (let i = 0; i < feeds.length; i++) {
-          if (feeds[i] && feeds[i].id === id) {
-            feeds[i].talking = false;
-          }
-        }
-        this.setState({feeds});
-      } else if (event === "destroyed") {
-        // The room has been destroyed
-        Janus.warn("The room has been destroyed!");
-      } else if (event === "event") {
-        if (msg["configured"] === "ok") {
-          // User published own feed successfully.
-          const user = {
-            ...this.state.user,
-            extra: {
-              ...(this.state.user.extra || {}),
-              streams: msg.streams,
-            },
-          };
-          this.setState({user});
-          if (this.state.muteOtherCams) {
-            this.setState({videos: NO_VIDEO_OPTION_VALUE});
-            this.state.virtualStreamingJanus.setVideo(NO_VIDEO_OPTION_VALUE);
-          }
-        } else if (msg["publishers"] !== undefined && msg["publishers"] !== null) {
-          // User just joined the room.
-          const feeds = sortAndFilterFeeds(msg["publishers"].filter((l) => (l.display = JSON.parse(l.display))));
-          Janus.debug("New list of available publishers/feeds:", this.state.feeds, feeds);
-          this.makeSubscription(feeds);
-        } else if (msg["leaving"] !== undefined && msg["leaving"] !== null) {
-          // One of the publishers has gone away?
-          const leaving = msg["leaving"];
-          Janus.log("Publisher left: " + leaving);
-          this.unsubscribeFrom([leaving], /* onlyVideo= */ false);
-        } else if (msg["unpublished"] !== undefined && msg["unpublished"] !== null) {
-          const unpublished = msg["unpublished"];
-          Janus.log("Publisher left: " + unpublished);
-          if (unpublished === "ok") {
-            // That's us
-            videoroom.hangup();
-            return;
-          }
-          this.unsubscribeFrom([unpublished], /* onlyVideo= */ false);
-        } else if (msg["error"] !== undefined && msg["error"] !== null) {
-          if (msg["error_code"] === 426) {
-            Janus.log("This is a no such room");
-          } else {
-            Janus.log(msg["error"]);
-          }
-        }
+          error: (error) => {
+            console.log("Error destroying VirtualStreaming on exit room", error);
+          },
+        });
       }
-    }
-    if (jsep !== undefined && jsep !== null) {
-      Janus.debug("Handling SDP as well...");
-      Janus.debug(jsep);
-      videoroom.handleRemoteJsep({jsep: jsep});
-    }
+      if (!reconnect && isFullScreen()) {
+        toggleFullScreen();
+      }
+      setTimeout(() => {
+        // if (videoroom) videoroom.detach();
+        // if (janus) janus.destroy();
+        if (!reconnect) {
+          this.state.virtualStreamingJanus.muteAudioElement();
+        } else {
+          this.state.virtualStreamingJanus.unmuteAudioElement();
+        }
+        this.setState({
+          muted: false,
+          question: false,
+          feeds: [],
+          mids: [],
+          localAudioTrack: null,
+          localVideoTrack: null,
+          upval: null,
+          remoteFeed: null,
+          videoroom: null,
+          subscriber: null,
+          janus: null,
+          delay: reconnect,
+          room: reconnect ? room : "",
+          chatMessagesCount: 0,
+          isSettings: false,
+        });
+        if (typeof callback === "function") callback();
+      }, 2000);
+    })
+
   };
 
-  newRemoteFeed = (subscription) => {
-    this.state.janus.attach({
-      plugin: "janus.plugin.videoroom",
-      opaqueId: "remotefeed_user",
-      success: (pluginHandle) => {
-        const remoteFeed = pluginHandle;
-        Janus.log(
-          `2 Plugin attached! (${remoteFeed.getPlugin()}, id=${remoteFeed.getId()}). -- This is a multistream subscriber ${remoteFeed}`
-        );
-        this.setState({remoteFeed, creatingFeed: false});
-        // We wait for the plugin to send us an offer
-        const subscribe = {request: "join", room: this.state.room, ptype: "subscriber", streams: subscription};
-        remoteFeed.send({message: subscribe});
-      },
-      error: (error) => {
-        Janus.error("  -- Error attaching plugin...", error);
-      },
-      iceState: (state) => {
-        Janus.log("ICE state (remote feed) changed to " + state);
-      },
-      webrtcState: (on) => {
-        Janus.log("Janus says this WebRTC PeerConnection (remote feed) is " + (on ? "up" : "down") + " now");
-      },
-      slowLink: (uplink, nacks) => {
-        Janus.warn(
-          "Janus reports problems " +
-            (uplink ? "sending" : "receiving") +
-            " packets on this PeerConnection (remote feed, " +
-            nacks +
-            " NACKs/s " +
-            (uplink ? "received" : "sent") +
-            ")"
-        );
-      },
-      onmessage: (msg, jsep) => {
-        const event = msg["videoroom"];
-        Janus.log(`::: Got a message (subscriber) ::: Event: ${event} Msg: `, msg);
-        if (msg["error"] !== undefined && msg["error"] !== null) {
-          Janus.debug("-- ERROR: " + msg["error"]);
-        } else if (event !== undefined && event !== null) {
-          if (event === "attached") {
-            this.setState({creatingFeed: false});
-            Janus.log("Successfully attached to feed in room " + msg["room"]);
-          } else if (event === "event") {
-            // Check if we got an event on a simulcast-related event from this publisher
-          } else {
-            // What has just happened?
-          }
-        }
-        if (msg["streams"]) {
-          // Update map of subscriptions by mid
-          const mids = Object.assign([], this.state.mids);
-          for (let i in msg["streams"]) {
-            let mindex = msg["streams"][i]["mid"];
-            //let feed_id = msg["streams"][i]["feed_id"];
-            mids[mindex] = msg["streams"][i];
-          }
-          this.setState({mids});
-        }
-        if (jsep !== undefined && jsep !== null) {
-          const {remoteFeed} = this.state;
-          Janus.debug("Handling SDP as well...");
-          Janus.debug(jsep);
-          // Answer and attach
-          remoteFeed.createAnswer({
-            jsep: jsep,
-            // Add data:true here if you want to subscribe to datachannels as well
-            // (obviously only works if the publisher offered them in the first place)
-            media: {audioSend: false, videoSend: false, data: false}, // We want recvonly audio/video
-            success: (jsep) => {
-              Janus.debug("Got SDP!");
-              Janus.debug(jsep);
-              let body = {request: "start", room: this.state.room};
-              remoteFeed.send({message: body, jsep: jsep});
-            },
-            error: (error) => {
-              Janus.error("WebRTC error:", error);
-              Janus.debug("WebRTC error... " + JSON.stringify(error));
-            },
-          });
-        }
-      },
-      onlocaltrack: (track, on) => {
-        // The subscriber stream is recvonly, we don't expect anything here
-      },
-      onremotetrack: (track, mid, on) => {
-        Janus.log(" ::: Got a remote track event ::: (remote feed)");
-        if (!mid) {
-          mid = track.id.split("janus")[1];
-        }
-        Janus.log("Remote track (mid=" + mid + ") " + (on ? "added" : "removed") + ":", track);
-        // Which publisher are we getting on this mid?
-        let {mids} = this.state;
-        let feed = mids[mid].feed_id;
-        Janus.log(" >> This track is coming from feed " + feed + ":", mid);
-        if (on) {
-          // If we're here, a new track was added
-          if (track.kind === "audio") {
-            // New audio track: create a stream out of it, and use a hidden <audio> element
-            let stream = new MediaStream();
-            stream.addTrack(track.clone());
-            Janus.log("Created remote audio stream:", stream);
-            let remoteaudio = this.refs["remoteAudio" + feed];
-            Janus.attachMediaStream(remoteaudio, stream);
-          } else if (track.kind === "video" && !track.muted) {
-            const remotevideo = this.refs["remoteVideo" + feed];
-            // New video track: create a stream out of it
-            const stream = new MediaStream();
-            stream.addTrack(track.clone());
-            Janus.log("Created remote video stream:", stream);
-            Janus.attachMediaStream(remotevideo, stream);
-          }
-        }
-      },
-      ondataopen: (label) => {
-        Janus.log("Feed - DataChannel is available! (" + label + ")");
-      },
-      ondata: (data, label) => {
-        Janus.debug("Feed - Got data from the DataChannel! (" + label + ")" + data);
-      },
-      ondataerror: (error) => {
-        Janus.warn("Feed - DataChannel error: " + error);
-      },
-      oncleanup: () => {
-        Janus.log(" ::: Got a cleanup notification (remote feed) :::");
-      },
-    });
-  };
+  // publishOwnFeed = (useVideo, useAudio) => {
+  //   const {videoroom, media} = this.state;
+  //   const {
+  //     audio: {audio_device},
+  //     video: {setting, video_device},
+  //   } = media;
+  //   const offer = {audioRecv: false, videoRecv: false, audioSend: useAudio, videoSend: useVideo, data: false};
+  //
+  //   if (useVideo) {
+  //     const {width, height, ideal} = setting;
+  //     offer.video = {width, height, frameRate: {ideal, min: 1}, deviceId: {exact: video_device}};
+  //   }
+  //
+  //   if (useAudio) {
+  //     offer.audio = {noiseSuppression: true, deviceId: {exact: audio_device}};
+  //   }
+  //
+  //   videoroom.createOffer({
+  //     media: offer,
+  //     simulcast: false,
+  //     success: (jsep) => {
+  //       Janus.debug("Got publisher SDP!");
+  //       Janus.debug(jsep);
+  //       const publish = {request: "configure", audio: useAudio, video: useVideo, data: false};
+  //       videoroom.send({message: publish, jsep: jsep});
+  //       if (!useVideo) {
+  //         media.video?.stream?.getTracks().forEach((t) => t.stop());
+  //         this.setState({cammuted: true});
+  //       }
+  //     },
+  //     error: (error) => {
+  //       Janus.error("WebRTC error:", error);
+  //     },
+  //   });
+  // };
+  //
+  // iceRestart = () => {
+  //   const {videoroom, remoteFeed} = this.state;
+  //
+  //   if (videoroom) {
+  //     videoroom.createOffer({
+  //       media: {audioRecv: false, videoRecv: false, audioSend: true, videoSend: true},
+  //       iceRestart: true,
+  //       simulcast: false,
+  //       success: (jsep) => {
+  //         Janus.debug("Got publisher SDP!");
+  //         Janus.debug(jsep);
+  //         const publish = {request: "configure", restart: true};
+  //         videoroom.send({message: publish, jsep: jsep});
+  //       },
+  //       error: (err) => {
+  //         Janus.error("WebRTC error:", err);
+  //       },
+  //     });
+  //   }
+  //
+  //   if (remoteFeed) remoteFeed.send({message: {request: "configure", restart: true}});
+  //   if (this.state.virtualStreamingJanus) this.state.virtualStreamingJanus.iceRestart();
+  //   iceRestartKliOlami();
+  // };
 
-  // Subscribe to feeds, whether already existing in the room, when I joined
-  // or new feeds that join the room when I'm already in. In both cases I
-  // should add those feeds to my feeds list.
-  // In case of feeds just joined and |question| is set, we should notify the
-  // new entering user by notifying everyone.
-  // Subscribes selectively to different stream types |subscribeToVideo|, |subscribeToAudio|, |subscribeToData|.
-  // This is required to stop and then start only the videos to save bandwidth.
   makeSubscription = (newFeeds) => {
     console.log("makeSubscription", newFeeds);
     const subscription = [];
@@ -1205,9 +921,7 @@ class VirtualClient extends Component {
   subscribeTo = (subscription) => {
     // New feeds are available, do we need create a new plugin handle first?
     if (this.state.remoteFeed) {
-      this.state.remoteFeed.send({
-        message: {request: "subscribe", streams: subscription},
-      });
+      this.state.subscriber.sub(subscription);
       return;
     }
 
@@ -1222,15 +936,74 @@ class VirtualClient extends Component {
 
     // We don't creating, so let's do it
     this.setState({creatingFeed: true});
-    this.newRemoteFeed(subscription);
+
+    // We wait for the plugin to send us an offer
+    this.state.subscriber.join(subscription, this.state.room).then(data => {
+      console.log('[client] Subscriber join: ', data)
+
+      this.onUpdateStreams(data.streams);
+
+      this.setState({remoteFeed: true, creatingFeed: false});
+    });
+
   };
+
+  handleTalking = (id, talking) => {
+    const feeds = Object.assign([], this.state.feeds);
+    for (let i = 0; i < feeds.length; i++) {
+      if (feeds[i] && feeds[i].id === id) {
+        feeds[i].talking = talking;
+      }
+    }
+    this.setState({feeds});
+  }
+
+  onUpdateStreams = (streams) => {
+    const mids = Object.assign([], this.state.mids);
+    for (let i in streams) {
+      let mindex = streams[i]["mid"];
+      //let feed_id = streams[i]["feed_id"];
+      mids[mindex] = streams[i];
+    }
+    this.setState({mids});
+  }
+
+  onRemoteTrack = (track, mid, on) => {
+    console.log(" ::: Got a remote track event ::: (remote feed)");
+    if (!mid) {
+    mid = track.id.split("janus")[1];
+  }
+    console.log("Remote track (mid=" + mid + ") " + (on ? "added" : "removed") + ":", track);
+  // Which publisher are we getting on this mid?
+  let {mids} = this.state;
+  let feed = mids[mid].feed_id;
+      console.log(" >> This track is coming from feed " + feed + ":", mid);
+  if (on) {
+    // If we're here, a new track was added
+    if (track.kind === "audio") {
+      // New audio track: create a stream out of it, and use a hidden <audio> element
+      let stream = new MediaStream();
+      stream.addTrack(track.clone());
+      console.log("Created remote audio stream:", stream);
+      let remoteaudio = this.refs["remoteAudio" + feed];
+      remoteaudio.srcObject = stream;
+    } else if (track.kind === "video") {
+      const remotevideo = this.refs["remoteVideo" + feed];
+      // New video track: create a stream out of it
+      const stream = new MediaStream();
+      stream.addTrack(track.clone());
+      console.log("Created remote video stream:", stream);
+      remotevideo.srcObject = stream;
+    }
+  }
+}
 
   // Unsubscribe from feeds defined by |ids| (with all streams) and remove it when |onlyVideo| is false.
   // If |onlyVideo| is true, will unsubscribe only from video stream of those specific feeds, keeping those feeds.
   unsubscribeFrom = (ids, onlyVideo) => {
     const {feeds} = this.state;
     const idsSet = new Set(ids);
-    const unsubscribe = {request: "unsubscribe", streams: []};
+    const streams = [];
     feeds
       .filter((feed) => idsSet.has(feed.id))
       .forEach((feed) => {
@@ -1240,17 +1013,17 @@ class VirtualClient extends Component {
           feed.streams
             .filter((stream) => stream.type === "video")
             .map((stream) => ({feed: feed.id, mid: stream.mid}))
-            .forEach((stream) => unsubscribe.streams.push(stream));
+            .forEach((stream) => streams.push(stream));
         } else {
           // Unsubscribe the whole feed (all it's streams).
-          unsubscribe.streams.push({feed: feed.id});
-          Janus.log("Feed " + JSON.stringify(feed) + " (" + feed.id + ") has left the room, detaching");
+          streams.push({feed: feed.id});
+          console.log("Feed " + JSON.stringify(feed) + " (" + feed.id + ") has left the room, detaching");
         }
       });
     // Send an unsubscribe request.
     const {remoteFeed} = this.state;
-    if (remoteFeed !== null && unsubscribe.streams.length > 0) {
-      remoteFeed.send({message: unsubscribe});
+    if (remoteFeed !== null && streams.length > 0) {
+      this.state.subscriber.unsub(streams);
     }
     if (!onlyVideo) {
       this.setState({feeds: feeds.filter((feed) => !idsSet.has(feed.id))});
@@ -1321,7 +1094,7 @@ class VirtualClient extends Component {
 
   sendKeepAlive = () => {
     const {user, janus} = this.state;
-    if (user && janus && janus.isConnected() && user.session && user.handle) {
+    if (user && janus && janus.isConnected && user.session && user.handle) {
       api
         .updateUser(user.id, user)
         .then((data) => {
@@ -1402,11 +1175,7 @@ class VirtualClient extends Component {
   };
 
   camMute = (cammuted) => {
-    const {videoroom, media} = this.state;
-    const {
-      video: {setting, video_device},
-    } = media;
-    const {width, height, ideal} = setting;
+    const {videoroom} = this.state;
 
     const user = Object.assign({}, this.state.user);
     if (user.role === userRolesEnum.ghost) return;
@@ -1414,36 +1183,9 @@ class VirtualClient extends Component {
 
     if (videoroom) {
       if (!cammuted) {
-        videoroom.createOffer({
-          media: {removeVideo: true},
-          simulcast: false,
-          success: (jsep) => {
-            Janus.debug("Got publisher SDP!");
-            Janus.debug(jsep);
-            videoroom.send({message: {request: "configure"}, jsep: jsep});
-          },
-          error: (error) => {
-            Janus.error("WebRTC error:", error);
-          },
-        });
-        this.stopLocalMedia();
+        this.stopLocalMedia(videoroom);
       } else {
-        videoroom.createOffer({
-          media: {
-            addVideo: true,
-            video: {width, height, frameRate: {ideal, min: 1}, deviceId: {exact: video_device}},
-          },
-          simulcast: false,
-          success: (jsep) => {
-            Janus.debug("Got publisher SDP!");
-            Janus.debug(jsep);
-            videoroom.send({message: {request: "configure"}, jsep: jsep});
-          },
-          error: (error) => {
-            Janus.error("WebRTC error:", error);
-          },
-        });
-        this.startLocalMedia();
+        this.startLocalMedia(videoroom);
       }
 
       user.camera = cammuted;
@@ -1463,36 +1205,36 @@ class VirtualClient extends Component {
     }
   };
 
-  stopLocalMedia = () => {
-    const {
-      media: {video},
-      cammuted,
-    } = this.state;
+  stopLocalMedia = (videoroom) => {
+    const {media, cammuted} = this.state;
     if (cammuted) return;
     console.log("Stop local video stream");
-    video?.stream?.getTracks().forEach((t) => t.stop());
-    this.setState({cammuted: true});
+    media.video?.stream?.getTracks().forEach((t) => t.stop());
+    //media.video.stream = null;
+    this.setState({cammuted: true, media});
+    if(videoroom) videoroom.mute(true)
   };
 
-  startLocalMedia = () => {
-    const {
-      media: {video: {devices, video_device} = {}},
-      cammuted,
-    } = this.state;
+  startLocalMedia = (videoroom) => {
+    const {media: {video: {devices, video_device} = {}}, cammuted,} = this.state;
     if (!cammuted) return;
     console.log("Bind local video stream");
     const deviceId = video_device || devices?.[0]?.deviceId;
     if (deviceId) {
       this.setVideoDevice(deviceId, true).then(() => {
+        const {stream} = this.state.media.video
+        if(videoroom) videoroom.mute(false, stream)
         this.setState({cammuted: false});
       });
     }
   };
 
   micMute = () => {
-    const {videoroom, muted} = this.state;
-    muted ? videoroom.unmuteAudio() : videoroom.muteAudio();
-    this.setState({muted: !muted});
+    const {media: {audio: {stream}}, muted} = this.state;
+    if(stream) {
+      stream.getAudioTracks()[0].enabled = muted;
+      this.setState({muted: !muted});
+    }
   };
 
   otherCamsMuteToggle = () => {
@@ -1531,13 +1273,14 @@ class VirtualClient extends Component {
         },
       });
     } else {
+      const {ip, country} = user;
       virtualStreamingJanus.init(user);
       stateUpdate.sourceLoading = true;
       this.setState(stateUpdate);
     }
   };
 
-  toggleKliOlami = (isKliOlamiShown = !this.state.isKliOlamiShown) => this.setState({isKliOlamiShown});
+  toggleQuad = (isKliOlamiShown = !this.state.isKliOlamiShown) => this.setState({isKliOlamiShown});
 
   updateLayout = (currentLayout) => {
     this.setState({currentLayout}, () => {
@@ -1737,7 +1480,7 @@ class VirtualClient extends Component {
             className={classNames("bottom-toolbar__item")}
             disableElevation
           >
-            <Mute t={t} action={this.micMute.bind(this)} disabled={!localAudioTrack} isOn={muted} ref="canvas1"/>
+            <Mute t={t} action={this.micMute.bind(this)} isOn={muted} ref="canvas1"/>
             <MuteVideo
               t={t}
               action={this.camMute.bind(this)}
@@ -1750,7 +1493,7 @@ class VirtualClient extends Component {
 
           <ButtonGroup className={classNames("bottom-toolbar__item")} variant="contained" disableElevation>
             <Fullscreen t={t} isOn={isFullScreen()} action={toggleFullScreen}/>
-            <KliOlamiToggle isOn={isKliOlamiShown} action={this.toggleKliOlami}/>
+            <KliOlamiToggle isOn={isKliOlamiShown} action={this.toggleQuad}/>
             <CloseBroadcast
               t={t}
               isOn={shidur}
@@ -1776,7 +1519,7 @@ class VirtualClient extends Component {
             <AskQuestion
               t={t}
               isOn={!!question}
-              disabled={!mqttOn || premodStatus || !audio_device || !localAudioTrack || delay || otherFeedHasQuestion}
+              disabled={!mqttOn || premodStatus || !audio_device || delay || otherFeedHasQuestion}
               action={this.handleQuestion.bind(this)}
             />
             <Vote t={t} id={user?.id} disabled={!user || !user.id || room === ""}/>
@@ -1940,7 +1683,7 @@ class VirtualClient extends Component {
           {/* ---------- */}
           <ButtonGroup
             variant="outlined"
-            disableElevation
+            disableElervation
             className={classNames("top-toolbar__item", "top-toolbar__toggle")}
           >
             <Badge color="secondary" badgeContent={asideMsgCounter.chat} showZero={false}>
@@ -2058,8 +1801,8 @@ class VirtualClient extends Component {
     }
 
     const kliOlami = !sourceLoading && isKliOlamiShown && (
-      <KliOlamiStream
-        close={() => this.toggleKliOlami(false)}
+      <QuadStream JanusStream={this.state.virtualStreamingJanus}
+        close={() => this.toggleQuad(false)}
         toggleAttach={(val = !kliOlamiAttached) => this.setState({kliOlamiAttached: val})}
         attached={kliOlamiAttached}
       />
@@ -2691,7 +2434,7 @@ class VirtualClient extends Component {
   }
 }
 
-const WrappedClass = withTranslation()(withTheme(VirtualClient));
+const WrappedClass = withTranslation()(withTheme(VirtualMqttClient));
 
 export default class WrapperForThemes extends React.Component {
   render() {
